youtube-project/channel-monitor/requirements.txt
google-auth==2.41.1
google-auth-oauthlib==1.2.2
google-auth-httplib2==0.2.0
google-api-python-client==2.183.0
isodate==0.7.2
streamlit==1.38.0
plotly==5.24.1
pandas==2.2.3
